+++ src/store/index.js
@@ -2,4 +2,5 @@ export * from './_MenuListState';
 export * from './_OrderState';
 export * from './_OrderListState';
 export * from './MenuListState';
+export * from './FilterListState';
 export * from './key';
