+++ src/store/key.js
@@ -3,5 +3,6 @@ export const KEY = {
   NEXT_ORDER_ID: 'nexrOrderId',
   ORDER: 'orderState',
   ORDER_SUM_PRICE: 'orderSumPrice',
-  MENU_LIST: 'menuListState'
+  MENU_LIST: 'menuListState',
+  FILTER_LIST: 'filterListState',
 }
